Added rules to properly ignore dnp3test and testset
Linux and Windows handle case sensitivity differently.  Namely, the
former thinks that "dnp3test" and "DNP3Test" are different, while the
latter believes that they are (mistakenly) the same.  Since we generate
binaries that match directories in a case insensitive manner
(dnp3test->DNP3Test and testset->TestSet), simply adding 'dnp3test' and
'testset' to the .gitignore causes the source code directories with the
"same" name to also be ignored.  This isn't what we want, so the trick
is to tell git to ignore the binaries but include the directories.  We
do this using the "negate" operator and the "directory-slash" suffix.
These rules are at the top of the .gitignore so that all the other rules
(like ignoring object files, swap files, etc.) behave like one would
expect.